feat: Moves common process_host_module to util.
